HTML, mejora de diseño

+++ HTML/Scripts/Eliminar/eliminar-api.php
@@ -1,6 +1,10 @@
 <?php
 	include "../variables.php";
-
+    if (!$_SESSION["login"]) {
+        header('Location: ../../index.php');
+		exit;
+    }
+	
 	$id_api = $_POST['id_api'];
 	$sql_delapi = "delete from api where id = $id_api";
     if ($conexion->query($sql_delapi) == True) {
@@ -10,5 +14,7 @@
 		$_SESSION["error"] = True;
 		$_SESSION["info"] = "API no se pudo eliminar.";
 	}
+
 	header("Location: ../../index.php");
+	exit;
 ?>
