+++ HTML/Scripts/Eliminar/eliminar-rol.php
@@ -1,6 +1,9 @@
 <?php
 	include "../variables.php";
-
+    if (!$_SESSION["login"]) {
+        header('Location: ../../index.php');
+		exit;
+    }
 	$id_rol = $_POST['id_rol'];
 	$sql_delrol = "delete from roles where id = $id_rol";
     if ($conexion->query($sql_delrol) == True) {
@@ -12,4 +15,5 @@
 	}
 	
 	header("Location: ../../index.php");
+	exit;
 ?>
